+++ Color.h
@@ -3,13 +3,15 @@
 #include <iostream>
 #include <SFML/Graphics.hpp>
 #include "functions.h"
+#include <stdint.h>
 
 class Color
 { 
 public:
-    int r_, g_, b_;
+    uint8_t r_, g_, b_;
+    
     Color(double r_rel, double g_rel, double b_rel);
-    Color(int r, int g, int b); 
+    Color(uint8_t r, uint8_t g, uint8_t b); 
     Color(sf::Color color);
     Color(){};
 
@@ -19,12 +21,12 @@ public:
 
     void claim();
     
-    void set_color(const int &r, const int &g, const int &b);
+    void set_color(const uint8_t &r, const uint8_t &g, const uint8_t &b);
     void set_relation(const double &r_rel, const double &g_rel, const double &b_rel);
         
-    void set_color_r(const int &r);
-    void set_color_g(const int &g);
-    void set_color_b(const int &b);
+    void set_color_r(const uint8_t &r);
+    void set_color_g(const uint8_t &g);
+    void set_color_b(const uint8_t &b);
 
     void set_relation_r(const double &r_rel);
     void set_relation_g(const double &g_rel);
@@ -45,24 +47,24 @@ public:
 
     Color &operator =(const Color &that) = default;    
 
-    friend Color operator + (const Color &color1,  const Color &color2);
-    friend Color operator + (const Color &color,   const int &number);
-    friend Color operator + (const int &number,    const Color &color);
-    friend Color operator + (const Color &color,   const double &number);
-    friend Color operator + (const double &number, const Color &color);
+    friend Color operator + (const Color &color1,   const Color &color2);
+    friend Color operator + (const Color &color,    const uint8_t &number);
+    friend Color operator + (const uint8_t &number, const Color &color);
+    friend Color operator + (const Color &color,    const double &number);
+    friend Color operator + (const double &number,  const Color &color);
 
     friend Color operator - (const Color &color1, const Color &color2);
-    friend Color operator - (const Color &color,  const int &number);
+    friend Color operator - (const Color &color,  const uint8_t &number);
     friend Color operator - (const Color &color,  const double &number);
 
     friend Color operator * (const Color &color1, const Color &color2);
-    friend Color operator * (const Color &color,  const int &number);
-    friend Color operator * (const int &number,   const Color &color);
+    friend Color operator * (const Color &color,  const uint8_t &number);
+    friend Color operator * (const uint8_t &number,   const Color &color);
     friend Color operator * (const Color &color,  const double &number);
     friend Color operator * (const double &number,const Color &color);
 
     friend Color operator / (const Color &color1, const Color &color2);
-    friend Color operator / (const Color &color,  const int &number);
+    friend Color operator / (const Color &color,  const uint8_t &number);
     friend Color operator / (const Color &color,  const double &number);
 
 private:
@@ -72,12 +74,12 @@ private:
 
 namespace Colors
 {
-    static const Color Black(0, 0, 0);       
-    static const Color White(255, 255, 255);       
-    static const Color Red(255, 0, 0);         
-    static const Color Green(0, 255, 0);       
-    static const Color Blue(0, 0, 255);        
-    static const Color Yellow(255, 255, 0);      
-    static const Color Magenta(255, 0, 255);     
-    static const Color Cyan(0, 255, 255);        
+    static const Color Black  ((uint8_t)0,   (uint8_t)0,   (uint8_t)0);       
+    static const Color White  ((uint8_t)255, (uint8_t)255, (uint8_t)255);       
+    static const Color Red    ((uint8_t)255, (uint8_t)0,   (uint8_t)0);         
+    static const Color Green  ((uint8_t)0,   (uint8_t)255, (uint8_t)0);       
+    static const Color Blue   ((uint8_t)0,   (uint8_t)0,   (uint8_t)255);        
+    static const Color Yellow ((uint8_t)255, (uint8_t)255, (uint8_t)0);      
+    static const Color Magenta((uint8_t)255, (uint8_t)0,   (uint8_t)255);     
+    static const Color Cyan   ((uint8_t)0,   (uint8_t)255, (uint8_t)255);        
 }
